requirements.txt
requests
PyYAML
Pillow